TRUNK-4747 Changes following review - move database change set to right folder - remove some comments
