+++ api/src/main/java/org/openmrs/util/PrivilegeConstants.java
@@ -881,9 +881,7 @@ public class PrivilegeConstants {
 	@AddOnStartup(description = "Able to manage the search index")
 	public static final String MANAGE_SEARCH_INDEX = "Manage Search Index";
 
-	/* allergy-api */
 	@AddOnStartup(description = "Able to add, edit, delete allergies")
 	public static final String MODIFY_ALLERGIES = "Task: Modify Allergies";
-	/* allergy-api */
 
 }
